chore: Update NominatedContract component
